Adding new intermediate module: analytics-data-processors
