+++ components/analytics/analytics-data-connectors/org.wso2.carbon.analytics.datasource.rdbms/src/main/java/org/wso2/carbon/analytics/datasource/rdbms/RDBMSUtils.java
