+++ components/analytics/analytics-data-connectors/org.wso2.carbon.analytics.datasource.rdbms/src/test/java/org/wso2/carbon/analytics/datasource/rdbms/H2FileDBAnalyticsFileSystemTest.java
@@ -18,8 +18,8 @@
  */
 package org.wso2.carbon.analytics.datasource.rdbms;
 
-import java.sql.Connection;
-import java.sql.SQLException;
+import java.io.File;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -29,64 +29,56 @@ import javax.naming.NamingException;
 import org.apache.tomcat.jdbc.pool.DataSource;
 import org.apache.tomcat.jdbc.pool.PoolProperties;
 import org.testng.annotations.BeforeSuite;
-import org.testng.annotations.Parameters;
-import org.wso2.carbon.analytics.datasource.core.AnalyticsRecordStore;
-import org.wso2.carbon.analytics.datasource.core.AnalyticsRecordStoreTest;
 import org.wso2.carbon.analytics.datasource.core.AnalyticsException;
-import org.wso2.carbon.analytics.datasource.rdbms.RDBMSQueryConfigurationEntry;
-import org.wso2.carbon.analytics.datasource.rdbms.RDBMSAnalyticsRecordStore;
+import org.wso2.carbon.analytics.datasource.core.AnalyticsFileSystem;
+import org.wso2.carbon.analytics.datasource.core.AnalyticsFileSystemTest;
 
 /**
- * MySQL implementation of analytics file system tests.
+ * H2 implementation of analytics file system tests.
  */
-public class MySQLInnoDBAnalyticsFileSystemTest extends AnalyticsRecordStoreTest {
+public class H2FileDBAnalyticsFileSystemTest extends AnalyticsFileSystemTest {
 
     @BeforeSuite
-    @Parameters({"mysql.url", "mysql.username", "mysql.password"})
-    public void setup(String url, String username, 
-            String password) throws NamingException, AnalyticsException, SQLException {
-        this.initDS(url, username, password);
-        AnalyticsRecordStore afs = new RDBMSAnalyticsRecordStore(this.generateQueryConfiguration());
-        Map<String, String> props = new HashMap<String, String>();
-        props.put("datasource", "DS");
-        afs.init(props);
-        this.init("MySQLInnoDBAnalyticsDataSource", afs);
+    public void setup() throws NamingException, AnalyticsException, IOException {
+        AnalyticsFileSystem afs = cleanupAndCreateAFS();
+        this.init("H2FileDBAnalyticsDataSource", afs);
     }
     
-    private void initDS(String url, String username, String password) throws NamingException, SQLException {
+    public static AnalyticsFileSystem cleanupAndCreateAFS() throws NamingException, IOException, AnalyticsException {
+        String dbPath = System.getProperty("java.io.tmpdir") + File.separator + "bam_test_db";
+        deleteFile(dbPath + ".mv.db");
+        deleteFile(dbPath + ".trace.db");
+        initDS("jdbc:h2:" + dbPath, "wso2carbon", "wso2carbon");
+        AnalyticsFileSystem afs = new RDBMSAnalyticsFileSystem(generateQueryConfiguration());
+        Map<String, String> props = new HashMap<String, String>();
+        props.put("datasource", "DSFS");
+        afs.init(props);
+        return afs;
+    }
+    
+    private static void deleteFile(String path) {
+        new File(path).delete();
+    }
+    
+    private static void initDS(String url, String username, String password) throws NamingException {
         PoolProperties pps = new PoolProperties();
-        pps.setDriverClassName("com.mysql.jdbc.Driver");
+        pps.setDriverClassName("org.h2.Driver");
         pps.setUrl(url);
         pps.setUsername(username);
         pps.setPassword(password);
+        pps.setDefaultAutoCommit(false);
         DataSource dsx = new DataSource(pps);
-        new InitialContext().bind("DS", dsx);
-        this.dropSystemTables(dsx);
+        new InitialContext().bind("DSFS", dsx);
     }
     
-    private void dropSystemTables(javax.sql.DataSource ds) throws SQLException {
-        Connection conn = null;
-        try {
-            conn = ds.getConnection();
-            conn.prepareStatement("DROP TABLE IF EXISTS AN_FS_DATA").executeUpdate();
-            conn.prepareStatement("DROP TABLE IF EXISTS AN_FS_PATH").executeUpdate();
-        } finally {
-            if (conn != null) {
-                try {
-                    conn.close();
-                } catch (SQLException ignore) { } 
-            }
-        }
-    }
-    
-    private RDBMSQueryConfigurationEntry generateQueryConfiguration() {
+    private static RDBMSQueryConfigurationEntry generateQueryConfiguration() {
         RDBMSQueryConfigurationEntry conf = new RDBMSQueryConfigurationEntry();
         String[] fsTableInitQueries = new String[3];
         fsTableInitQueries[0] = "CREATE TABLE AN_FS_PATH (path VARCHAR(256), is_directory BOOLEAN, length BIGINT, parent_path VARCHAR(256), PRIMARY KEY(path), FOREIGN KEY (parent_path) REFERENCES AN_FS_PATH(path) ON DELETE CASCADE)";
         fsTableInitQueries[1] = "CREATE TABLE AN_FS_DATA (path VARCHAR(256), sequence BIGINT, data BLOB, PRIMARY KEY (path,sequence), FOREIGN KEY (path) REFERENCES AN_FS_PATH(path) ON DELETE CASCADE)";
         fsTableInitQueries[2] = "CREATE INDEX index_parent_id ON AN_FS_PATH(parent_path)";        
         conf.setFsTableInitQueries(fsTableInitQueries);        
-        conf.setFsTablesCheckQuery("SELECT record_id FROM AN_FS_PATH WHERE path = '/'");
+        conf.setFsTablesCheckQuery("SELECT path FROM AN_FS_PATH WHERE path = '/'");
         conf.setFsPathRetrievalQuery("SELECT * FROM AN_FS_PATH WHERE path = ?");
         conf.setFsListFilesQuery("SELECT path FROM AN_FS_PATH WHERE parent_path = ?");
         conf.setFsInsertPathQuery("INSERT INTO AN_FS_PATH (path,is_directory,length,parent_path) VALUES (?,?,?,?)");
