+++ components/analytics/analytics-data-connectors/org.wso2.carbon.analytics.datasource.rdbms/src/test/java/org/wso2/carbon/analytics/datasource/rdbms/MySQLInnoDBAnalyticsFileSystemTest.java
@@ -18,7 +18,8 @@
  */
 package org.wso2.carbon.analytics.datasource.rdbms;
 
-import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -28,34 +29,52 @@ import javax.naming.NamingException;
 import org.apache.tomcat.jdbc.pool.DataSource;
 import org.apache.tomcat.jdbc.pool.PoolProperties;
 import org.testng.annotations.BeforeSuite;
-import org.wso2.carbon.analytics.datasource.core.AnalyticsFileSystem;
-import org.wso2.carbon.analytics.datasource.core.AnalyticsFileSystemTest;
+import org.testng.annotations.Parameters;
+import org.wso2.carbon.analytics.datasource.core.AnalyticsRecordStore;
+import org.wso2.carbon.analytics.datasource.core.AnalyticsRecordStoreTest;
 import org.wso2.carbon.analytics.datasource.core.AnalyticsException;
-import org.wso2.carbon.analytics.datasource.rdbms.RDBMSQueryConfigurationEntry;
 
 /**
- * H2 implementation of analytics data source tests.
+ * MySQL implementation of analytics file system tests.
  */
-public class H2MemDBAnalyticsFileSystemTest extends AnalyticsFileSystemTest {
+public class MySQLInnoDBAnalyticsFileSystemTest extends AnalyticsRecordStoreTest {
 
     @BeforeSuite
-    public void setup() throws NamingException, AnalyticsException, IOException {
-        this.initDS("jdbc:h2:mem:bam_test_db", "wso2carbon", "wso2carbon");
-        AnalyticsFileSystem afs = new RDBMSAnalyticsFileSystem(this.generateQueryConfiguration());
+    @Parameters({"mysql.url", "mysql.username", "mysql.password"})
+    public void setup(String url, String username, 
+            String password) throws NamingException, AnalyticsException, SQLException {
+        this.initDS(url, username, password);
+        AnalyticsRecordStore afs = new RDBMSAnalyticsRecordStore(this.generateQueryConfiguration());
         Map<String, String> props = new HashMap<String, String>();
         props.put("datasource", "DS");
         afs.init(props);
-        this.init("H2MemDBAnalyticsDataSource", afs);
+        this.init("MySQLInnoDBAnalyticsDataSource", afs);
     }
     
-    private void initDS(String url, String username, String password) throws NamingException {
+    private void initDS(String url, String username, String password) throws NamingException, SQLException {
         PoolProperties pps = new PoolProperties();
-        pps.setDriverClassName("org.h2.Driver");
+        pps.setDriverClassName("com.mysql.jdbc.Driver");
         pps.setUrl(url);
         pps.setUsername(username);
         pps.setPassword(password);
         DataSource dsx = new DataSource(pps);
         new InitialContext().bind("DS", dsx);
+        this.dropSystemTables(dsx);
+    }
+    
+    private void dropSystemTables(javax.sql.DataSource ds) throws SQLException {
+        Connection conn = null;
+        try {
+            conn = ds.getConnection();
+            conn.prepareStatement("DROP TABLE IF EXISTS AN_FS_DATA").executeUpdate();
+            conn.prepareStatement("DROP TABLE IF EXISTS AN_FS_PATH").executeUpdate();
+        } finally {
+            if (conn != null) {
+                try {
+                    conn.close();
+                } catch (SQLException ignore) { } 
+            }
+        }
     }
     
     private RDBMSQueryConfigurationEntry generateQueryConfiguration() {
@@ -65,7 +84,7 @@ public class H2MemDBAnalyticsFileSystemTest extends AnalyticsFileSystemTest {
         fsTableInitQueries[1] = "CREATE TABLE AN_FS_DATA (path VARCHAR(256), sequence BIGINT, data BLOB, PRIMARY KEY (path,sequence), FOREIGN KEY (path) REFERENCES AN_FS_PATH(path) ON DELETE CASCADE)";
         fsTableInitQueries[2] = "CREATE INDEX index_parent_id ON AN_FS_PATH(parent_path)";        
         conf.setFsTableInitQueries(fsTableInitQueries);        
-        conf.setFsTablesCheckQuery("SELECT path FROM AN_FS_PATH WHERE path = '/'");
+        conf.setFsTablesCheckQuery("SELECT record_id FROM AN_FS_PATH WHERE path = '/'");
         conf.setFsPathRetrievalQuery("SELECT * FROM AN_FS_PATH WHERE path = ?");
         conf.setFsListFilesQuery("SELECT path FROM AN_FS_PATH WHERE parent_path = ?");
         conf.setFsInsertPathQuery("INSERT INTO AN_FS_PATH (path,is_directory,length,parent_path) VALUES (?,?,?,?)");
